// packages/contracts/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        AiSuggestRequest: {
            p_session: string;
            p_recipient: string;
            p_age: string;
            p_occasion: string;
            p_tone: string;
            /** Format: int64 */
            p_budget: number | string;
        };
// ...
        FlowResult: {
            result: null | components["schemas"]["JsonElement"];
        };
// ...
        GiftSurveyResponse: {
            /** Format: uuid */
            id: string;
            source: string;
            suggestions: components["schemas"]["GiftSuggestionResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}